+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Last Changes
 
+- [#23](https://github.com/LaxarJS/shop-demo/issues/23): project: simplified the RequireJS configuration
 - [#21](https://github.com/LaxarJS/shop-demo/issues/21): use laxar-testing instead of laxar/laxar_testing
     + NEW FEATURE: see ticket for details
 - [#22](https://github.com/LaxarJS/shop-demo/issues/22): project: switched to flow-based grunt tasks