+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Last Changes
 
+- [#31](https://github.com/LaxarJS/shop-demo/issues/31): project: use release versions of laxarjs artifacts
 - [#29](https://github.com/LaxarJS/shop-demo/issues/29): documentation: use yeoman rather than grunt-init
 - [#30](https://github.com/LaxarJS/shop-demo/issues/30): article-teaser-widget: small fixes
 - [#27](https://github.com/LaxarJS/shop-demo/issues/27): project: re-implemented article-teaser-widget using React