+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Last Changes
 
+- [#32](https://github.com/LaxarJS/shop-demo/issues/32): documentation: explain React-based widget in tutorial
 - [#31](https://github.com/LaxarJS/shop-demo/issues/31): project: use release versions of laxarjs artifacts
 - [#29](https://github.com/LaxarJS/shop-demo/issues/29): documentation: use yeoman rather than grunt-init
 - [#30](https://github.com/LaxarJS/shop-demo/issues/30): article-teaser-widget: small fixes